+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 0.1.1 (Next)
 
+* [#3](https://github.com/dblock/danger-toc/issues/3): Configure "Table of Contents" header with `toc.header` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2017/11/5)